refactor(playback): add feature barrel; route immersive imports through barrels (dep:check)
